Melhorias na logica e limpeza das interações entre service controller e repository
